backend/requirements.txt
alabaster==1.0.0
annotated-types==0.7.0
anyio==4.12.1
argcomplete==3.6.3
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgiref==3.11.0
asttokens==3.0.1
async-lru==2.1.0
attrs==25.4.0
babel==2.18.0
beautifulsoup4==4.14.3
bleach==6.3.0
blivet==3.12.1
Brlapi==0.8.7
certifi==2025.7.9
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.7
cockpit @ file:///builddir/build/BUILD/cockpit-356-build/cockpit-356/tmp/wheel/cockpit-356-py3-none-any.whl#sha256=5a84a26ba9e7f611c00c6efb3a3e675039514318856e1355bd01b72f1d86f01c
comm==0.2.3
contourpy==1.3.3
crypt_r==3.13.1
cryptography==46.0.5
cupshelpers==1.0
cycler==0.12.1
dasbus==1.7
dbus-python==1.4.0
dbus_next==0.2.3
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.1
distro==1.9.0
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnf==4.24.0
docutils==0.22.4
et_xmlfile==2.0.0
executing==2.2.1
fastjsonschema==2.21.2
fedora-third-party==0.10
fica==0.4.1
file-magic==0.4.0
fonttools==4.61.1
fqdn==1.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-genai==1.65.0
googleapis-common-protos==1.72.0
greenlet==3.3.1
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
humanize==4.13.0
idna==3.10
imagesize==1.4.1
ipykernel==7.1.0
ipylab==1.1.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
iso639==0.1.4
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
jupytext==1.19.1
kiwisolver==1.4.9
langtable==0.0.69
libcomps==0.1.23
libdnf==0.75.0
libdnf5==5.2.18.0
louis==3.33.0
lxml==6.0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdit-py-plugins==0.5.0
mdurl==0.1.2
mistune==3.2.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
nftables==0.1
notebook==7.5.3
notebook_shim==0.2.4
numpy==2.4.2
olefile==0.47
openpyxl==3.1.5
otter-grader==6.1.6
packaging==25.0
pandas==3.0.0
pandocfilters==1.5.1
parso==0.8.5
perf==0.1
pexpect==4.9.0
pid==3.0.4
pillow==12.1.0
platformdirs==4.5.1
playwright==1.58.0
ply==3.11
productmd==1.50
prometheus_client==0.24.1
prompt_toolkit==3.0.52
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pwquality==1.4.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
PyAudio==0.2.13
pycairo==1.28.0
pycparser==2.22
pycups==2.0.4
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
pyenchant==3.2.2
Pygments==2.19.2
PyGObject==3.54.5
pyinotify==0.9.6
PyJWT==2.10.1
pykickstart @ file:///builddir/build/BUILD/pykickstart-3.66-build/pykickstart-3.66
pyparsing==3.3.2
pyparted==3.13.0
PySocks==1.7.1
python-augeas==1.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.2
python-json-logger==4.0.0
python-linux-procfs==0.7.3
python-meh==0.52
python-on-whales==0.80.0
python-pam==2.0.2
pyudev==0.24.3
pyxdg==0.28
PyYAML==6.0.2
pyynl @ file:///builddir/build/BUILD/kernel-6.18.13-build/kernel-6.18.13/linux-6.18.13-200.fc43.x86_64/tools/net/ynl
pyzmq==27.1.0
qrcode==8.2
RapidFuzz==3.12.2
referencing==0.36.2
regex==2025.10.23
requests==2.32.5
requests-file==2.0.0
requests-ftp==0.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roman-numerals==4.1.0
rpds-py==0.27.0
rpm==6.0.1
scipy==1.17.0
selinux @ file:///builddir/build/BUILD/libselinux-3.9-build/libselinux-3.9/src
Send2Trash==2.1.0
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.9-build/selinux-3.9/python/sepolicy
setools==4.6.0
setuptools==78.1.1
simpleline==1.9.0
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
sos==4.11.0
soupsieve==2.8
Sphinx==9.1.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.4
stack-data==0.6.3
systemd-python==235
tenacity==9.1.4
terminado==0.18.1
tinycss2==1.4.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.6.3
wcwidth==0.5.3
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
widgetsnbextension==4.0.15
wrapt==1.17.3
xkbregistry==0.3
